475: add funnel.master.js to handle worker
