+++ .eslintrc
@@ -15,7 +15,7 @@
           "object": true
         },
         "AssignmentExpression": {
-          "array": true,
+          "array": false,
           "object": false
         }
       },
