+++ src/plan/test/funnel.spec.js
@@ -5,8 +5,7 @@ import { Point } from '../../global/class/Point';
 import { Polypoint } from '../../interpret/class/Polypoint';
 import { Triangle } from '../../interpret/class/Triangle';
 import { PolypointState } from '../class/PolypointState';
-import { funnelPolypointsFromPortals } from '../funnel.worker';
-import { getPortals } from '../portals';
+import { getPortals, funnelPolypointsFromPortals } from '../funnel';
 import { backwardsFunnelPath, backwardsFunnelGetClearancePoint } from './backwardsFunnelData';
