+++ src/think/fsm.js
@@ -27,7 +27,7 @@ import { getMyRole } from '../look/playerRoles';
 import { Point } from '../global/class/Point';
 import { getDTGraph } from '../interpret/interpret';
 import { getShortestPolypointPath } from '../plan/astar';
-import { funnelPolypoints } from '../plan/funnel';
+import { funnelEnemyPolypoints } from '../plan/funnel.master';
 
 
 let stateMessage = ''; // a short description of the bot's current state and next action
@@ -76,13 +76,17 @@ function chaseEnemy(me, enemy) {
   // Runtime: O(M*CPTL^2) with visualizations on, O(M + S*CPTL^2) with visualizations off
   const enemyShortestPath = [];
   _.forEach(
-    funnelPolypoints(getShortestPolypointPath(
+    funnelEnemyPolypoints(getShortestPolypointPath(
       getPlayerCenter(enemy),
       getEnemyGoal(),
       getDTGraph(),
     ).shortestPath, getDTGraph()),
     polypoint => enemyShortestPath.push(polypoint),
   );
+
+  if (_.isEmpty(enemyShortestPath)) {
+    return { globalGoal: findAllyFlagStation(), enemyShortestPath: null };
+  }
 
   // Set goal as the interception point
   const interceptionPolypoint = _.find(getPointsAlongPath(enemyShortestPath), polypoint => {
